do not round too early
git-svn-id: https://svn.r-project.org/R/trunk@59106 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/unix/sys-unix.c
@@ -208,13 +208,13 @@ void R_getProcTime(double *data)
     getrusage(RUSAGE_SELF, &self);
     getrusage(RUSAGE_CHILDREN, &children);
     data[0] = (double) self.ru_utime.tv_sec +
-	1e-3 * ((double) self.ru_utime.tv_usec/1000);
+	1e-3 * (self.ru_utime.tv_usec/1000);
     data[1] = (double) self.ru_stime.tv_sec +
-	1e-3 * ((double) self.ru_stime.tv_usec/1000);
+	1e-3 * ((self.ru_stime.tv_usec/1000);
     data[3] = (double) children.ru_utime.tv_sec +
-	1e-3 * ((double) children.ru_utime.tv_usec/1000);
+	1e-3 * (children.ru_utime.tv_usec/1000);
     data[4] = (double) children.ru_stime.tv_sec +
-	1e-3 * ((double) children.ru_stime.tv_usec/1000);
+	1e-3 * (children.ru_stime.tv_usec/1000);
 #else
     /* Not known to be currently used */
     struct tms timeinfo;
